+++ examples/cases/tpsl/create_position_tpsl_order.py
@@ -8,8 +8,7 @@ from examples.utils import (
     get_adjust_price_by_pct,
 )
 from x10.config import BTC_USD_MARKET
-from x10.perpetual.order_object import OrderTpslTriggerParam, create_order_object
-from x10.perpetual.orders import (
+from x10.models.order import (
     OrderPriceType,
     OrderSide,
     OrderTpslType,
@@ -17,6 +16,7 @@ from x10.perpetual.orders import (
     OrderType,
     TimeInForce,
 )
+from x10.perpetual.order_object import OrderTpslTriggerParam, create_order_object
 
 LOGGER = logging.getLogger()
 MARKET_NAME = BTC_USD_MARKET
